+++ src/ResolvedDomainName.php
@@ -9,7 +9,7 @@ interface ResolvedDomainName extends Host, ExternalDomainName
     /**
      * Returns the domain effective tld component.
      */
-    public function publicSuffix(): EffectiveTLD;
+    public function suffix(): EffectiveTLD;
 
     /**
      * Returns the second level domain component.
@@ -19,7 +19,7 @@ interface ResolvedDomainName extends Host, ExternalDomainName
     /**
      * Returns the registrable domain component.
      */
-    public function registrableDomain(): self;
+    public function registrableDomain(): DomainName;
 
     /**
      * Returns the sub domain component.
@@ -56,5 +56,5 @@ interface ResolvedDomainName extends Host, ExternalDomainName
      * If the domain already has a public suffix it will be replaced by the new value
      * otherwise the public suffix content is added to or remove from the current domain.
      */
-    public function withPublicSuffix(EffectiveTLD $publicSuffix): self;
+    public function withSuffix(EffectiveTLD $publicSuffix): self;
 }